+++ Assets/Scripts/Scriptables/Systems/OrbitAroundOwnerSystem.cs
@@ -23,7 +23,7 @@ namespace Assets.Scripts.Scriptables.Systems
 
 		public override SystemTargets Targets => SystemTargets.Projectile;
 
-		public override string GetTooltip(ushort level)
+		public override string GetTooltip(int level)
 		{
 			return "Causes an entity to orbit around its creator";
 		}
