+++ Assets/Scripts/Scriptables/Systems/SetLifetimeProportionalToChargeSystem.cs
@@ -19,7 +19,7 @@ namespace Assets.Scripts.Scriptables.Systems
 
 		public override SystemTargets Targets => SystemTargets.Projectile;
 
-		public override string GetTooltip(ushort level) => $"Lifetime increases by <color=\"green\">{multiplier * 100:0}%</color> with weapon charge.";
+		public override string GetTooltip(int level) => $"Lifetime increases by <color=\"green\">{multiplier * 100:0}%</color> with weapon charge.";
 
 		public void OnEntityEffectAdded(Entity entity, EffectWrapper wrapper)
 		{
